Create README
Project Proposal 
Topic: Graphical Tip Calculator

Project Abstraction
The purpose of the project is to create a graphical tip calculator that is used to any service render company or institution (such as hotels, restaurants, valet parking, etc.), when the user makes a service charge after he or she gets services from the service provider company, they should insert the amount of money into a text field and select the desired tip percentage using a button or a slider. Therefore, the program should include two buttons, one to calculate the tip when the user presses and the other one is to reset the button. In addition, When the user presses the calculate Tip button, the tip amount and the total amount are displayed in separate labels. 

Project Relevance
For the implementation of this project, the project will link with some of the educational goals of the classes such as Graphic User Interface, Object-Oriented Design, and Test-Driven Development.

Project Contribution
While I am doing this project, my contribution will be starting the projects from the scratch and building the code step by step.
Background
	The following URL that used to reference the project. 
Building
	For building the program I can use NetBeans IDE 8.2 or
	IntelliJ IDEA 2020.3.2

Running
	The program will include the main and other classes.
	To run simply select the run button to display the program.
Required Resources
	Team members should be experienced with Java.